feat: worker update v2
