+++ src/renderer/entities/proxy/index.ts
@@ -1,3 +1,2 @@
 export { proxyModel } from './model/proxy-model';
-export * from './lib/constants';
 export { ProxyAccount } from './ui/ProxyAccount/ProxyAccount';
