--- src/renderer/entities/proxy/lib/types.ts
@@ -1,3 +1,0 @@
-import type { AccountId, ProxyAccount } from '@shared/core';
-
-export type ProxyStore = Record<AccountId, ProxyAccount[]>;
