+++ src/renderer/entities/proxy/ui/ProxyAccount/ProxyAccount.tsx
@@ -3,7 +3,17 @@ import { BodyText, DropdownIconButton, HelpText, Identicon, Truncate } from '@sh
 import { AccountId, ProxyType } from '@shared/core';
 import { useI18n } from '@app/providers';
 import { DropdownIconButtonOption } from '@shared/ui/types';
-import { ProxyTypeName } from '../../lib/constants';
+
+const ProxyTypeName: Record<ProxyType, string> = {
+  [ProxyType.ANY]: 'proxy.names.any',
+  [ProxyType.NON_TRANSFER]: 'proxy.names.nonTransfer',
+  [ProxyType.STAKING]: 'proxy.names.staking',
+  [ProxyType.AUCTION]: 'proxy.names.auction',
+  [ProxyType.CANCEL_PROXY]: 'proxy.names.cancelProxy',
+  [ProxyType.GOVERNANCE]: 'proxy.names.governance',
+  [ProxyType.IDENTITY_JUDGEMENT]: 'proxy.names.identityJudgement',
+  [ProxyType.NOMINATION_POOLS]: 'proxy.names.nominationPools',
+};
 
 type Props = {
   className?: string;
