+++ src/renderer/features/proxies/lib/utils.ts
@@ -1,67 +1,17 @@
-import { u8aToHex } from '@polkadot/util';
-import { decodeAddress } from '@polkadot/util-crypto';
-import { WellKnownChain } from '@substrate/connect';
-
-import { AccountId, Chain, ChainId, ProxyAccount, Account, ProxiedAccount, ProxyAction, NoID } from '@shared/core';
-import { AccountType, NotificationType, Wallet } from '@shared/core';
+import { NotificationType, Chain } from '@shared/core';
+import type { ProxyAccount, Account, ProxyAction, NoID, Wallet } from '@shared/core';
 import { dictionary } from '@shared/lib/utils';
 
-export const proxyWorkerUtils = {
-  toAccountId,
+export const proxiesUtils = {
   isRegularProxy,
-  isSameProxy,
-  getKnownChain,
-  isProxiedAccount,
-  getNotification_NEW,
+  getNotification,
 };
-
-/**
- * Try to get account id of the address
- * WARNING! Duplication for worker
- * @param address account's address
- * @return {String}
- */
-export function toAccountId(address: string): AccountId {
-  try {
-    return u8aToHex(decodeAddress(address));
-  } catch {
-    return '0x00';
-  }
-}
 
 function isRegularProxy(chain: Chain): boolean {
   return Boolean(chain.options?.includes('regular_proxy'));
 }
 
-function isSameProxy(oldProxy: NoID<ProxyAccount>, newProxy: NoID<ProxyAccount>): boolean {
-  return (
-    oldProxy.accountId === newProxy.accountId &&
-    oldProxy.proxiedAccountId === newProxy.proxiedAccountId &&
-    oldProxy.chainId === newProxy.chainId &&
-    oldProxy.proxyType === newProxy.proxyType &&
-    oldProxy.delay === newProxy.delay
-  );
-}
-
-const enum Chains {
-  POLKADOT = '0x91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3',
-  KUSAMA = '0xb0a8d493285c2df73290dfb7e61f870f17b41801197a149ca93654499ea3dafe',
-}
-
-const KnownChains: Record<ChainId, WellKnownChain> = {
-  [Chains.POLKADOT]: WellKnownChain.polkadot,
-  [Chains.KUSAMA]: WellKnownChain.ksmcc3,
-};
-
-function getKnownChain(chainId: ChainId): WellKnownChain | undefined {
-  return KnownChains[chainId];
-}
-
-function isProxiedAccount(account: Pick<Account, 'type'>): account is ProxiedAccount {
-  return account.type === AccountType.PROXIED;
-}
-
-function getNotification_NEW(
+function getNotification(
   proxies: ProxyAccount[],
   wallets: Wallet[],
   accounts: Account[],
